Implemented tree inverse dynamics solver and generic (with arbitrary base) kinematic and dynamics loops

+++ src/chainidsolver_recursive_newton_euler_floating_base.cpp
@@ -96,7 +96,7 @@ namespace KDL{
         }
         base_force = X[0]*f[0];
         //debug
-        for(int i=0; i < ns; i++) {
+        for(int i=0; i < (int)ns; i++) {
             Segment segm;
         	segm = chain.getSegment(i);
             //std::cout << "bLink " << segm.getName() << " a= " << a[i] << " f= " << f[i] << std::endl;
